Changed Triangles Function
For Export, uncomment:
16, 379, 381-386

Changed Triangles function to be two separate parts- Top and Bottom
Triangle.

Changed Triangles to be oriented in the direction of the average of each
corner's normal vector, and each triangle to offset in the Z direction
by the average of the depth of each corner. Can't get Helix to work in
order to check it it.

and there is  boarder that I'm not sure how to get rid of